Load tiny_mce only when needed, and as late as possible
This change ensures it won't be loaded when it's not needed, and
attempts to defer its loading as long as it can. It does not appear
to play nice with LabJS at the moment, so it is still loaded directly
from a script-tag, but the tag is pushed to the end of the document
to ensure that it doesn't block rendering.

Also drops useless loading of YUI components.