Retry removals that fail
This partly resolves removal of packages that other packages depend on.
Ie. package A depends on B, but B was first on the command-line, so it
will fail to be removed. Now it'll let B fail, then remove A, then
retry B.

This won't resolve large dependency trees, but will work for the
majority of packages (most of which will only have 1-2 layers of
dependencies).